requirements.txt
paho-mqtt
BeautifulSoup
pylibconfig2
